opengoo_1.2_beta

+++ fengoffice_/application/models/PublicFiles.class.php
@@ -67,7 +67,11 @@
     */
     static function deleteFile($delete_file) {
       $destination = self::getFilePath($delete_file);
-      return @unlink($destination);
+      if (is_file($destination)) {
+      	return @unlink($destination);
+      } else {
+      	return false;
+      }
     } // deleteFile
     
     /**
